refactor(redis): RedisStoreBuilder
* fix(redis): refund balance if settlement fails

* refactor(redis): separate test modules into files

also move helpers into tests/common
